feat: Tour management UI, Map integration & Backend CRUD
Implementiran je kompletan interfejs za kreiranje tura sa Leaflet mapama (CreateTourView) i javni katalog tura (ToursView).

Dodata je kontrolna tabla za vodiče (MyToursView) koja omogućava upravljanje životnim ciklusom ture (Draft/Publish/Archive/Delete) povezivanjem na ažurirani Pinia store.

Ažurirani su backend servisi (TourService) i kontroleri da podrže kompletan CRUD i ispravljeno je rutiranje za promenu statusa prebacivanjem na POST metodu.

+++ services/Tours.API/Application/Services/CartService.cs
@@ -13,7 +13,16 @@ public class CartService : ICartService
 
     public async Task<Guid> AddItemAsync(Guid userId, AddCartDto dto)
     {
-        var i = new CartItem { Id = Guid.NewGuid(), UserId = userId, TourId = dto.TourId, Name = dto.Name, Price = dto.Price };
+        var i = new CartItem
+        {
+            Id = Guid.NewGuid(),
+            UserId = userId,
+            TourId = dto.TourId,
+            Name = dto.Name,
+            // FIX: Dodato (decimal) jer je dto.Price double
+            Price = (decimal)dto.Price
+        };
+
         await _ctx.Cart.InsertOneAsync(i);
         return i.Id;
     }
